T47: extract RecordingsTableUI from ParsekUI
Move recordings table (2,286 lines, 57 fields, 30+ methods) into its
own RecordingsTableUI class in UI/. This is the largest extraction from
ParsekUI, covering window management, group/chain tree rendering,
sorting, rename/double-click handling, loop period editing, confirmation
dialogs, status/phase styles, and tooltip rendering.

ParsekUI keeps shared static loop helpers (used by SettingsWindowUI),
wipe confirmation dialogs, and thin forwarding wrappers so existing test
references (ParsekUI.FormatDuration, ParsekUI.SortColumn, etc.)
continue to work without test modifications.

GroupPickerUI ownership moves to RecordingsTableUI. New Flight and
InFlightMode accessors on ParsekUI enable recordings table to access
flight state for watch/rewind/FF buttons.
